+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+
+* [#395](https://github.com/babylonlabs-io/vigilante/pull/395) fix: remove mutex on pk in slasher
+
 ## v0.23.8
 
 ### Improvements